// WADNR.Web/src/app/shared/generated/model/geographic-assignment-step.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GeographicLookupItem } from './geographic-lookup-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class GeographicAssignmentStep { 
    ProjectID?: number;
    SelectedIDs?: Array<number> | null;
    NoSelectionExplanation?: string | null;
    AvailableOptions?: Array<GeographicLookupItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface GeographicAssignmentStepForm { 
    ProjectID?: FormControl<number>;
    SelectedIDs?: FormControl<Array<number>>;
    NoSelectionExplanation?: FormControl<string>;
    AvailableOptions?: FormControl<Array<GeographicLookupItem>>;
}

export class GeographicAssignmentStepFormControls { 
    public static ProjectID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SelectedIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static NoSelectionExplanation = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AvailableOptions = (value: FormControlState<Array<GeographicLookupItem>> | Array<GeographicLookupItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<GeographicLookupItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}