// WADNR.Web/src/app/shared/generated/model/geographic-override-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class GeographicOverrideRequest { 
    SelectedIDs?: Array<number> | null;
    NoSelectionExplanation?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface GeographicOverrideRequestForm { 
    SelectedIDs?: FormControl<Array<number>>;
    NoSelectionExplanation?: FormControl<string>;
}

export class GeographicOverrideRequestFormControls { 
    public static SelectedIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static NoSelectionExplanation = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}