// WADNR.Web/src/app/shared/generated/model/geographic-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class GeographicLookupItem { 
    ID?: number;
    DisplayName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface GeographicLookupItemForm { 
    ID?: FormControl<number>;
    DisplayName?: FormControl<string>;
}

export class GeographicLookupItemFormControls { 
    public static ID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static DisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}